Tehtävä 4.21

+++ osa4/controllers/blogs.js
@@ -32,8 +32,17 @@ blogsRouter.post("", async (request, response) => {
 })
 
 blogsRouter.delete("/:id", async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: "token invalid" }).end()
+    }
+    const blog = await Blog.findById(request.params.id)
+    if (blog.user.toString() === decodedToken.id.toString()) {
+        await Blog.findByIdAndDelete(request.params.id)
+        response.status(204).end()
+    } else {
+        response.status(401).json({ error: "user is not creator" }).end()
+    }
 })
 
 blogsRouter.put("/:id", async (request, response) => {
